Fix #3205: VE crash for objects with length exceeding 32-bit
Use 64-bit integers for number of elements / rows / columns, and the corresponding indices.

Guard against objects with extremely large extents in one or more dimensions.

+++ src/Host/API/Impl/RHostSession.Interop.cs
@@ -175,8 +175,8 @@ namespace Microsoft.R.Host.Client {
 
         private class RObjectInfo: IRObjectInformation {
             public string TypeName { get; set; }
-            public int Length { get; set; }
-            public IReadOnlyList<int> Dim { get; set; }
+            public long Length { get; set; }
+            public IReadOnlyList<long> Dim { get; set; }
         }
     }
 }
