+++ src/Package/Impl/DataInspect/VisualGrid/HeaderTextVisual.cs
@@ -29,11 +29,11 @@ namespace Microsoft.VisualStudio.R.Package.DataInspect {
         }
         #endregion
 
-        public HeaderTextVisual(int columnIndex) {
+        public HeaderTextVisual(long columnIndex) {
             ColumnIndex = columnIndex;
         }
 
-        public int ColumnIndex { get; }
+        public long ColumnIndex { get; }
 
         /// <summary>
         /// Name of the column
